+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 2.2.0 - 2015-03-01
+### Added
+- Optional "listen" flag for `@InjectPreference` annotation as shorthand for also adding `@OnPreferenceChange` annotation to the same element
+- `OnPreferenceChange` annotations on methods can now contain multiple keys (if the method has zero parameters)
+
+### Changed
+- Method calls now always occur after field updates for any given preference key (for both initialization and updates)
+- Annotated methods can now contain zero parameters, to disregard the preference value.
+
 ## 2.1.1 - 2015-02-25
 ### Added
 - `PreferenceInjector.inject(Fragment target)` method. The `Context` can be inferred from the `Fragment`.